Version Tracking and Display
Added  a mechanism to use a version number and build info that is
common to all executables. GripMMI shows the version number in the menu
bar and full build info in an ‘About …’ popup. The popup can be
displayed via the System Menu or by right clicking on the main window
background.

The latest build info is written to a text file when building
CreateStandaloneGripMMI. This file is visible to Git, making it easy to
follow which version and build corresponds to what commit.