Improve behavior of graph combo box.
The refresh timer was making it difficult to select a different
collection of graphs to show in the drop down combo box because the
screen would refresh too quickly. Here I try to inhibit the screen
refresh while the user is making a selection and then restart periodic
refresh.

NB There is still a bug. If the user pulls down the list but then
leaves the list by clicking elsewhere, the periodic refresh does not
get restarted.