Autoscaling index limits.
In the Views system it used to be that the Autoscale functions would
clip the min and max indices for the search to the min and max user X
value of the view. That assumes that the X axis will be the index into
the array, which is not true for XY and scatter plots. We now do the
search over the entire specified range of indices, independent of the
axis limits.