Bug Fix - InsertGripRealtimeDataInfo().
Fixed a bug in which realtime data from a GripRealtimeDataInfo
structure was not being properly inserted into an EPMTelemetry packet.

Because this routine was never used in a deliverable for the GripMMI
project, the error was not previously detected but had no effect. It
became important when CLWSemulator was enhanced to simulate more of the
GRIP realtime data.