Issue #16 Documentation of Emulator
RunGripMMI.bat has been modified to make it easier to configure
CLWSemulator.exe according to the new functionality. The comments have
been improved to show the emulator options.

The Installation instructions have also been modified to better
describe the emulator test options.

NB: The “Theory of Operation” chapter of the on-line wiki has also been
updated to describe options for testing with the CLWS emulator.