CLWS command line bug.
The CLWS simulator was not getting the correct command line arguments
when executing the .exe generated in release mode. This was because the
compiler option for the character set was configured for Unicode, even
though the debug configuration was correctly configured to ‘Not Set’.

The configuration notes were updated to included a specific caution
about making sure that both debug and release modes are configured
properly.

Some debugging output was added and at the same time both console
programs were changed to output text to stdout only, rather than a mix
of stdout and stderr.