Clean leaves GripExecutables intact.
It used to be that cleaning the solution would remove the executables
and batch file from GripExecutables and clean the cache directory. This
meant that we could not clean the solution of intermediate
compiler/linker files prior to a release. The makefiles in
CreateStandaloneGripMMI now leave the contents of GripExecutables
intact but they still clear out the cache directory. So it is advisable
to do a ‘Clean Solution’ in both the debug and release configurations
prior to packaging the project for release.