Issue #13 Fixed bug where TCP/IP freezes.
Previously the exchange of data via TCP/IP between CLWSemulator.exe and
GripGroundMonitorClient.exe stopped after about 1 hour.

This was the result of Alive packets not being read by
CLWSemulator.exe. Apparently, if packets are not recv()’d by
CLWSemulator.exe, buffers will fill up and the send() call on the
client side will block indefinitely. The solution that I have chosen is
to set a timeout for the send() calls. If such a timeout occurs,
further attempts to transmit the Alive packet are inhibited. This
solution is chosen over make CLWSemulator.exe recv() the Alive packets
because it will be more robust. I am not 100% certain that the real
CLWS server handles Alive packets. The solution implemented here will
keep the program from freezing even if the Alive packets are not read.