Improved script crawler manual entry.
Clicking on the text boxes to jump to a specific step could be
confusing because updates stop while the text box is active. The text
box is now highlighted when selected, as is the GoTo button to attract
the attention to the user about what needs to be done. Hitting <return>
will also act like clicking on the GoTo button.