Issue #14 Use RestartGripMMI to review data.
RunGripMMI.bat automatically generates a batch file, with name
RestartGripMMI.yyyy.mm.dd.bat, that can be used to restart the
graphical interface (GripMMI.exe) without restarting the TCP/IP client
(GripGroundMonitorClient.exe). This is useful for restarting the GUI
when the TCP/IP client is still running, but it can also be used to
start up the GUI on a pair of cache files that are already stored. So
these automatically generated batch files facilitate looking back at
data from previous sessions.