Improved handling when buffers full.
When the data buffers are saturated, GripMMI will give a single error
message that explains how to 1) continue viewing data that has already
be acquired or 2) restart in a way to monitor new incoming data.

Note that this should not happen in normal operation as the buffers can
hold 12 hours or more of continuous data transmissions.